Introduce CheckListContentWidget. Resuse it for CheckListPanel and and AcademicsHomePanel [#1714]
