+++ assets/strings.en.json
@@ -2547,6 +2547,8 @@
     "widget.gies.button.title.review": "Review Checklist",
     "widget.gies.button.title.view_notes": "View My Notes",
     "widget.gies.button.title.continue": "Continue",
+    "widget.checklist.gies.title": "iDegrees New Student Checklist",
+    "widget.checklist.uiuc.title": "New Student Checklist",
 
     "widgets.home.app_help.header.title": "App Help",
     "widgets.home.app_help.video_tutorial.button.title": "Video Tutorial",
