+++ lib/ui/academics/AcademicsHomePanel.dart
@@ -21,6 +21,7 @@ import 'package:illinois/service/FlexUI.dart';
 import 'package:illinois/service/Storage.dart';
 import 'package:illinois/ui/academics/AcademicsEventsContentWidget.dart';
 import 'package:illinois/ui/canvas/CanvasCoursesContentWidget.dart';
+import 'package:illinois/ui/gies/CheckListContentWidget.dart';
 import 'package:illinois/utils/AppUtils.dart';
 import 'package:rokwire_plugin/service/connectivity.dart';
 import 'package:rokwire_plugin/service/localization.dart';
@@ -228,11 +229,9 @@ class _AcademicsHomePanelState extends State<AcademicsHomePanel>
       case AcademicsContent.events:
         return AcademicsEventsContentWidget();
       case AcademicsContent.gies_checklist:
-        //TBD: DD - implement
-        return Container();
+        return CheckListContentWidget(contentKey: 'gies');
       case AcademicsContent.uiuc_checklist:
-        //TBD: DD - implement
-        return Container();
+        return CheckListContentWidget(contentKey: 'uiuc_student');
       case AcademicsContent.courses:
         return CanvasCoursesContentWidget();
       default:
